fuelgauge: add back battery level animation
This adds the battery level animation when entering battery
settings menu. This was dropped in AOSP, but still exists on Pixel's
(At least in DP4).

When a user enters the battery settings menu, the battery icon will
animate
into the current battery level, as well as the battery level text.

Updated for ODR3 (8.1) by @Sirhc95

Change-Id: Ibfe3cc50f192281a0bede670acf9ace1d7523f02